Fix some bugs

+++ src/main/java/seedu/address/model/student/dashboard/Progress.java
@@ -28,7 +28,7 @@ public class Progress {
 
         this.totalTasks = totalTasks;
         this.numCompletedTasks = numCompletedTasks;
-        progressValueInPercent = (totalTasks == 0) ? 0 : ((numCompletedTasks / totalTasks) * 100);
+        progressValueInPercent = (totalTasks == 0) ? 0 : (int) (((double) numCompletedTasks / totalTasks) * 100);
     }
 
     /**
